Windows: fix SectorAlignedWriter writing at EOF on pre-sized images
The Windows SectorAlignedWriter::new constructor seeded its tracked
`position` from file.metadata().len(), so a freshly-opened image that
had been pre-sized via set_len() (the re-resize restore path) put the
first write at EOF instead of offset 0. Sector 0 stayed zero-filled,
producing a restored disk with a zeroed partition table.

Seed `position` from the file's current cursor instead — fresh handles
and devices both start at 0, so the writer lands at sector 0 as the
restore code expects. Adds a cross-platform regression test that
constructs a writer over a set_len-grown tempfile and asserts byte 0
of the file changes after writing a sentinel.

Also clears a handful of unrelated build/clippy warnings the working
tree had picked up under newer toolchains (clippy 1.92) so the
pre-commit hook passes: unused CompilationResult in build.rs, a dead
std::io import in write_clonezilla_disk (the only `io::` reference
was inside a cfg(not(windows)) block), an unused RAII guard binding
in inspect_tab, a duplicated #![cfg(test)] inner attribute on the
qemu_img_test helper, redundant pointer casts and a redundant
closure in os::windows, two manual `% == 0` checks flagged by
clippy::manual_is_multiple_of, a non-test item appearing after the
test module in os::windows, two collapsible-else-if blocks in the
backup/restore tab cancel handlers, and two unnecessary u32 casts
in the test_clone_fix example.

+++ example/test_clone_fix.rs
@@ -45,11 +45,12 @@ fn main() {
         s.catalog_file_size / 1024
     );
 
-    let catalog_min = (s.catalog_file_size as u32)
+    let catalog_min = s
+        .catalog_file_size
         .saturating_mul(3)
         .saturating_div(2)
         .min(HFS_MAX_BTREE_FILE_SIZE);
-    let extents_min = (s.extents_file_size as u32).min(HFS_MAX_BTREE_FILE_SIZE);
+    let extents_min = s.extents_file_size.min(HFS_MAX_BTREE_FILE_SIZE);
 
     println!(
         "Building blank target file: {} MiB, bs={}, catalog>={} KiB, extents>={} KiB",
